+++ sclpl/cli/app.py
@@ -19,6 +19,7 @@ from sclpl.cli import (
     docs_cmd,
     launcher,
     plugin_cmd,
+    python_cmd,
     project_cmd,
     resource_cmd,
     test_cmd,
@@ -107,6 +108,7 @@ workflow_cmd.register(app)
 catalog_cmd.register(app)
 contract_cmd.register(app)
 plugin_cmd.register(app)
+python_cmd.register(app)
 project_cmd.register(app)
 resource_cmd.register(app)
 test_cmd.register(app)